requirements.txt
torch==2.4.1
torchvision==0.19.1
tqdm
open_clip_torch
ftfy
numpy
pandas
Pillow
regex
scipy
six
tensorboardX
tensorboardX